// src/interfaces/super_admin/subsuperadmindetails.interface.ts
export interface SubSuperAdminDetailsInt {
    _id: string;
    company_name: string;
    business_description: string;
    logo: string;
    mobile_no: number;
    superAdmin_id: object;
    is_active:number;
    branch: string; 
    expiry_on: Date;
    activated_on: Date;
    remark:string;
  }
  